Move Validator client to separate file
